# Webservice/server/swagger.yaml
openapi: 3.0.3
info:
  title: Travel Suite API
  description: |
    ## REST Web Service cho ứng dụng du lịch Travel Suite

    API hỗ trợ hai loại người dùng:
    - **Guide** – đăng địa điểm, upload ảnh, sửa/xóa địa điểm của mình
    - **Traveller** – tìm kiếm, xem thông tin, đánh giá và bình luận địa điểm

    ### Xác thực
    Sử dụng **JWT Bearer Token**. Sau khi đăng nhập / đăng ký, lưu `token` từ response và gửi theo header:
    ```
    Authorization: Bearer <token>
    ```

    ### Demo Accounts
    | Role | Email | Password |
    |------|-------|----------|
    | Guide | guide@travel.com | guide123 |
    | Traveller | traveller@travel.com | traveller123 |
  version: "1.0.0"
  contact:
    name: Travel Suite Support
servers:
  - url: http://localhost:3000/api
    description: Local Development Server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token nhận được sau khi đăng nhập
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error description"

    # ─── User ────────────────────────────────────────
...
    Place:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
        name:
          type: string
          example: "Hội An Ancient Town"
        description:
          type: string
          example: "A beautifully preserved trading port with lantern-lit streets..."
        location:
          type: string
          example: "Quảng Nam, Vietnam"
        category:
          type: string
          enum: [Heritage, Nature, Adventure, Beach, City, Mountain, Food, Culture, Other]
          example: "Heritage"
        guide_id:
          type: string
          format: uuid
        guide_name:
          type: string
          example: "guide_demo"
        avg_rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
          example: 4.3
        rating_count:
          type: integer
          example: 12
        cover_image:
          type: string
          nullable: true
        thumbnail:
          type: string
          nullable: true
          example: "abc123.jpg"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        total:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 12
        pages:
          type: integer
          example: 4

    # ─── Image ───────────────────────────────────────
...
    Comment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        place_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        username:
          type: string
          example: "traveller_demo"
        role:
          type: string
          enum: [guide, traveller]
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 4
        comment:
          type: string
          example: "Địa điểm rất đẹp và yên bình!"
        created_at:
          type: string
          format: date-time
    CommentStats:
      type: object
      properties:
        avg_rating:
          type: number
          format: float
          example: 4.2
        total:
          type: integer
          example: 15
paths:
...
  /places:
    get:
      tags: [Places]
      summary: Tìm kiếm / Lấy danh sách địa điểm
      description: |
        Lấy danh sách địa điểm với phân trang và bộ lọc.
        Hỗ trợ tìm kiếm theo từ khoá trong tên, mô tả, và địa chỉ.
        Không cần đăng nhập.
      operationId: getPlaces
      parameters:
        - name: q
          in: query
          description: Từ khoá tìm kiếm (tìm trong name, description, location)
          schema:
            type: string
            example: "Hội An"
        - name: category
          in: query
          description: Lọc theo danh mục
          schema:
            type: string
            enum: [Heritage, Nature, Adventure, Beach, City, Mountain, Food, Culture, Other]
        - name: guide_id
          in: query
          description: Lọc theo ID của guide (để xem các địa điểm của 1 guide cụ thể)
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Số trang (bắt đầu từ 1)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Số địa điểm mỗi trang
          schema:
            type: integer
            default: 12
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Danh sách địa điểm
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      places:
                        type: array
                        items:
                          $ref: '#/components/schemas/Place'
                      pagination:
                        $ref: '#/components/schemas/Pagination'

    post:
      tags: [Places]
      summary: Tạo địa điểm mới
      description: |
        Chỉ **Guide** mới có thể tạo địa điểm mới.
        Cần gửi JWT token trong header Authorization.
      operationId: createPlace
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, location, category]
              properties:
                name:
                  type: string
                  example: "Phú Quốc Island"
                description:
                  type: string
                  example: "Hòn đảo ngọc với bãi biển cát trắng và nước biển xanh trong..."
                location:
                  type: string
                  example: "Kiên Giang, Vietnam"
                category:
                  type: string
                  enum: [Heritage, Nature, Adventure, Beach, City, Mountain, Food, Culture, Other]
                  example: "Beach"
      responses:
        "201":
          description: Tạo địa điểm thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Place created"
                  data:
                    type: object
                    properties:
                      place:
                        $ref: '#/components/schemas/Place'
        "400":
          description: Thiếu trường bắt buộc hoặc category không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Chưa đăng nhập
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "403":
          description: Không phải Guide
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "Access restricted to: guide"
...
  /places/{id}/comments:
    get:
      tags: [Comments]
      summary: Lấy danh sách bình luận của địa điểm
      description: Trả về toàn bộ bình luận và thống kê đánh giá trung bình. Không cần đăng nhập.
      operationId: getComments
      parameters:
        - name: id
          in: path
          required: true
          description: UUID của địa điểm
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Danh sách bình luận và thống kê
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      comments:
                        type: array
                        items:
                          $ref: '#/components/schemas/Comment'
                      stats:
                        $ref: '#/components/schemas/CommentStats'
        "404":
          description: Không tìm thấy địa điểm
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags: [Comments]
      summary: Thêm bình luận và đánh giá
      description: |
        Chỉ **Traveller** mới có thể bình luận.
        Sau khi thêm bình luận, điểm đánh giá trung bình (`avg_rating`) của địa điểm sẽ được tự động cập nhật.
      operationId: addComment
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID của địa điểm
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating, comment]
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 4
                  description: Đánh giá từ 1 (tệ) đến 5 (tuyệt vời)
                comment:
                  type: string
                  minLength: 5
                  example: "Địa điểm rất đẹp, không gian yên tĩnh và văn hóa phong phú!"
      responses:
        "201":
          description: Thêm bình luận thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Comment added"
                  data:
                    type: object
                    properties:
                      comment:
                        $ref: '#/components/schemas/Comment'
                      avg_rating:
                        type: number
                        example: 4.2
        "400":
          description: Dữ liệu không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingFields:
                  value:
                    success: false
                    message: "Rating and comment required"
                invalidRating:
                  value:
                    success: false
                    message: "Rating must be 1–5"
                shortComment:
                  value:
                    success: false
                    message: "Comment must be at least 5 characters"
        "403":
          description: Chỉ Traveller mới có thể bình luận
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...